+++ src/main/java/com/smunity/server/domain/question/controller/QuestionController.java
@@ -4,8 +4,8 @@ import com.smunity.server.domain.question.dto.QuestionRequestDto;
 import com.smunity.server.domain.question.dto.QuestionResponseDto;
 import com.smunity.server.domain.question.service.QuestionCommandService;
 import com.smunity.server.domain.question.service.QuestionQueryService;
-import com.smunity.server.global.security.annotation.AuthAdmin;
-import com.smunity.server.global.security.annotation.AuthMember;
+import com.smunity.server.global.security.resolver.annotation.AuthAdmin;
+import com.smunity.server.global.security.resolver.annotation.AuthMember;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springdoc.core.annotations.ParameterObject;
