+++ src/main/java/com/smunity/server/global/config/WebConfig.java
@@ -1,7 +1,7 @@
 package com.smunity.server.global.config;
 
-import com.smunity.server.global.security.annotation.resolver.AuthAdminArgumentResolver;
-import com.smunity.server.global.security.annotation.resolver.AuthMemberArgumentResolver;
+import com.smunity.server.global.security.resolver.AuthAdminArgumentResolver;
+import com.smunity.server.global.security.resolver.AuthMemberArgumentResolver;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
